+++ inst/doc/family-level_taxonomy.R
@@ -1,5 +1,6 @@
 ## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
+  eval = SoilTaxonomy:::.soilDB_metadata_available(min_version = "2.0.0"),
   collapse = TRUE,
   comment = "#>"
 )
@@ -15,7 +16,8 @@ data("ST_family_classes")
 head(ST_family_classes)
 
 ## -----------------------------------------------------------------------------
-afamily_classes <- parse_family("fine-loamy, mixed, semiactive, mesic ultic haploxeralfs")$classes_split[[1]]
-class_descriptions <- subset(ST_family_classes, classname %in% afamily_classes)
-class_descriptions[match(afamily_classes, class_descriptions$classname),]
+f <- "fine-loamy, mixed, semiactive, mesic ultic haploxeralfs"
+family_classes <- parse_family(f, column_metadata = FALSE)$classes_split[[1]]
+
+get_ST_family_classes(classname = family_classes)